backend/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
python-multipart
sentence-transformers
faiss-cpu
transformers
torch
numpy
